+++ src/algorithms/operators/mod.rs
@@ -5,10 +5,12 @@
 // function re-exports in rustdoc.
 pub(crate) mod complementer;
 pub(crate) mod disjoint_union;
+pub(crate) mod intersection;
 pub(crate) mod simplify;
 pub(crate) mod union;
 
 pub use complementer::complementer;
 pub use disjoint_union::{disjoint_union, disjoint_union_many};
+pub use intersection::intersection;
 pub use simplify::simplify;
 pub use union::union;