+++ src/lib.rs
@@ -60,6 +60,7 @@ pub use crate::algorithms::connectivity::transitive_closure::transitive_closure;
 pub use crate::algorithms::io::edgelist::read_edgelist;
 pub use crate::algorithms::operators::complementer::complementer;
 pub use crate::algorithms::operators::disjoint_union::{disjoint_union, disjoint_union_many};
+pub use crate::algorithms::operators::intersection::intersection;
 pub use crate::algorithms::operators::simplify::simplify;
 pub use crate::algorithms::operators::union::union;
 pub use crate::algorithms::paths::dijkstra::dijkstra_distances;